requirements-pa.txt
# Core dependencies
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
gunicorn>=20.1.0
python-dotenv>=0.19.0
pydantic>=1.8.2
aiohttp>=3.8.0

# YouTube API
google-api-python-client>=2.0.0
google-auth-oauthlib>=0.4.6
youtube-transcript-api>=0.4.4

# Database
sqlalchemy>=1.4.23
aiosqlite>=0.17.0

# Machine Learning
torch>=2.0.0
transformers>=4.11.0

# WSGI server
waitress>=2.1.2

# Other
python-multipart>=0.0.5
